feature: 대략적인 쿼리 작성
