+++ server/config.ts
@@ -3,7 +3,7 @@ const mysql = require('mysql2');
 const mysqlssh = require('mysql-ssh');
 
 // create the connection to database
-const config = {
+export const config = {
   host: '3.36.99.206',
   port: '3306',
   user: 'danmin20',
